Creating a new (multiple classes) course
into an existing jupyterhub+nbgrader installation - generalized and updated version

You have to adopt some variables in the "__main__"-section according to your installation before running it.
This tools adds new groups and services to jupyterhub config file using append and update.
If the systemwide variables are adopted, only COURSE for the course_id and TEACHER for the teachers account is necessary.
If a course is created, you can also add students from a students csv listing.

updated to include teachers as jupyterhub admins